+++ src/Rule/Dictionary.php
@@ -92,7 +92,7 @@ final class Dictionary implements Rule
      */
     private function getDictionaryWord(string $password): ?string
     {
-        foreach ($this->getWordsToCheck($password) as $word) {
+        foreach ($this->getFormattedWords($password) as $word) {
             try {
                 if ($this->wordList->contains($word)) {
                     return $word;
@@ -106,32 +106,6 @@ final class Dictionary implements Rule
             }
         }
         return null;
-    }
-
-    /**
-     * @param string $word Word to check.
-     * @return Traversable<string> Variants of the word to check.
-     */
-    private function getWordsToCheck(string $word): Traversable
-    {
-        yield from $this->getUniqueWords($this->getFormattedWords($word));
-    }
-
-    /**
-     * @param iterable<string> $words Words to filter.
-     * @return Traversable<string> Unique words.
-     */
-    private function getUniqueWords(iterable $words): Traversable
-    {
-        $checked = [];
-        foreach ($words as $word) {
-            if (isset($checked[$word])) {
-                continue;
-            }
-
-            $checked[$word] = true;
-            yield $word;
-        }
     }
 
     /**